+++ unlock-app/src/components/interface/locks/Settings/forms/CreditCardCustomPrice.tsx
@@ -1,4 +1,10 @@
-import { Placeholder, ToggleSwitch, Input, Button } from '@unlock-protocol/ui'
+import {
+  Placeholder,
+  ToggleSwitch,
+  Input,
+  Button,
+  Select,
+} from '@unlock-protocol/ui'
 import React, { useEffect, useState } from 'react'
 import { useForm } from 'react-hook-form'
 import { ToastHelper } from '~/components/helpers/toast.helper'
@@ -11,9 +17,11 @@ import { SettingCardDetail } from '../elements/SettingCard'
 import { formatNumber } from '~/utils/formatter'
 import { storage } from '~/config/storage'
 import { useUSDPricing } from '~/hooks/useUSDPricing'
+import { currencies } from '@unlock-protocol/core'
 
 interface CreditCardFormSchema {
   creditCardPrice?: string | number | null
+  creditCardCurrency?: string
 }
 
 interface CreditCardCustomPriceProps {
@@ -31,6 +39,7 @@ export default function CreditCardCustomPrice({
 }: CreditCardCustomPriceProps) {
   const [useCustomPrice, setUseCustomPrice] = useState(false)
   const [hasPriceConversion, setHasPriceConversion] = useState(false)
+  const [currency, setCurrency] = useState('') // default currency is USD
 
   const {
     isLoading: isLoadingSettings,
@@ -39,18 +48,22 @@ export default function CreditCardCustomPrice({
 
   const getDefaultValues = async (): Promise<CreditCardFormSchema> => {
     const settings = (await storage.getLockSettings(network, lockAddress)).data
-    const price = settings?.creditCardPrice
+    const { creditCardPrice: price, creditCardCurrency } = settings
 
-    if (price) {
+    if (price && creditCardCurrency) {
       const priceInUsd = parseFloat(formatNumber(price / 100)).toFixed(2)
       setUseCustomPrice(true) // enable toggle because it has custom price
+      setCurrency(creditCardCurrency)
+      setValue('creditCardCurrency', `${creditCardCurrency}`)
       return {
         creditCardPrice: priceInUsd,
+        creditCardCurrency,
       }
     }
 
     return {
       creditCardPrice: undefined,
+      creditCardCurrency: 'usd',
     }
   }
 
@@ -58,7 +71,7 @@ export default function CreditCardCustomPrice({
     register,
     handleSubmit,
     setValue,
-    formState: { errors },
+    formState: { errors, isLoading: isLoadingForm },
   } = useForm<CreditCardFormSchema>({
     defaultValues: async () => await getDefaultValues(),
   })
@@ -83,6 +96,7 @@ export default function CreditCardCustomPrice({
 
   const onSaveCreditCardPrice = async ({
     creditCardPrice,
+    creditCardCurrency,
   }: CreditCardFormSchema) => {
     const price = creditCardPrice
       ? parseFloat(`${creditCardPrice}`) * 100 // save price in basis points
@@ -92,6 +106,7 @@ export default function CreditCardCustomPrice({
       lockAddress,
       network,
       creditCardPrice: price,
+      creditCardCurrency,
     })
 
     await ToastHelper.promise(savePricePromise, {
@@ -105,7 +120,7 @@ export default function CreditCardCustomPrice({
     onSaveCreditCardPrice(fields)
   }
 
-  if (isLoading || isLoadingSettings) {
+  if (isLoading || isLoadingSettings || isLoadingForm) {
     return (
       <Placeholder.Root>
         <Placeholder.Card />
@@ -118,6 +133,14 @@ export default function CreditCardCustomPrice({
     !hasPriceConversion || (useCustomPrice && hasPriceConversion)
 
   const symbol = lock?.currencySymbol
+
+  // list of currencies for dropdown
+  const currencyOptions = currencies.map(({ currency, symbol }) => {
+    return {
+      label: `${currency.toUpperCase()} - ${symbol}`,
+      value: currency,
+    }
+  })
 
   return (
     <div className="grid gap-2">
@@ -149,25 +172,43 @@ export default function CreditCardCustomPrice({
             }}
           />
           {toggleActive && (
-            <Input
-              type="number"
-              step="any"
-              disabled={disabled}
-              description="Set a fixed price in USD that is charged for card payments."
-              error={errors?.creditCardPrice?.message}
-              {...register('creditCardPrice', {
-                required: {
-                  value: toggleActive,
-                  message: 'This field is required.',
-                },
-                min: {
-                  value: CREDIT_CARD_MIN_USD_PRICE,
-                  message: `Price is too low for us to process credit cards. It needs to be at least ${parseFloat(
-                    `${CREDIT_CARD_MIN_USD_PRICE}`
-                  ).toFixed(2)}. `,
-                },
-              })}
-            />
+            <div className="grid gap-2 md:grid-cols-2">
+              <div className="flex flex-col gap-1">
+                <span>Price</span>
+                <Input
+                  type="number"
+                  step="any"
+                  disabled={disabled}
+                  description="Set a fixed price in USD that is charged for card payments."
+                  error={errors?.creditCardPrice?.message}
+                  {...register('creditCardPrice', {
+                    required: {
+                      value: toggleActive,
+                      message: 'This field is required.',
+                    },
+                    min: {
+                      value: CREDIT_CARD_MIN_USD_PRICE,
+                      message: `Price is too low for us to process credit cards. It needs to be at least ${parseFloat(
+                        `${CREDIT_CARD_MIN_USD_PRICE}`
+                      ).toFixed(2)}. `,
+                    },
+                  })}
+                />
+              </div>
+              <div className="flex flex-col gap-1">
+                <span>Currency</span>
+                <Select
+                  description="The selected currency will be used for credit card payments. (the currency is automatically updated when the selection is changed)"
+                  options={currencyOptions}
+                  disabled={disabled}
+                  defaultValue={currency}
+                  onChange={(currency) => {
+                    setCurrency(`${currency}`)
+                    setValue('creditCardCurrency', `${currency}`)
+                  }}
+                />
+              </div>
+            </div>
           )}
           <div className="w-full md:w-1/3">
             <Button className="w-full" size="small" disabled={saveDisabled}>
